requirements.txt
# requirements.txt - Versões estáveis e compatíveis com o ambiente
# Análise e correção definitiva por JarvisG

# Fixa a versão do Python para garantir a compatibilidade
# Nota: Se o seu app principal precisar do Python 3.11, você pode manter
# python-version==3.11
# Se não, 3.9 é uma escolha segura para compatibilidade geral.
# Por ora, vamos remover a linha para deixar o Streamlit usar a padrão estável.

# Bibliotecas principais do Streamlit e Langchain
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-groq
langchain-google-genai
python-dotenv
pypdf
beautifulsoup4
tiktoken
pyperclip
requests
faiss-cpu
pydub
langchain-huggingface
langchain-deepseek
fake-useragent
youtube-transcript-api
urllib3
langchain-playwright

# As bibliotecas abaixo não são mais necessárias com a nova abordagem
# openai-whisper
# ffmpeg-python
# yt-dlp

#.\venv\Scripts\Activate.ps1
#pip install -r requirements.txt
#cd C:\ProjetoOraculo
#pip install -r requirements.txt
#streamlit run Oraculo.py